+++ src/gf-form-locator/classes/class-form-locations-table.php
@@ -29,14 +29,11 @@ class Form_Locations_Table extends WP_List_Table {
 	}
 
 	/**
-	 * Retrieve customer’s data from the database
+	 * Retrieve customer’s data from the database.
 	 *
-	 * @param int $per_page     The number of entries to list on a page.
-	 * @param int $page_number  The page number.
-	 *
-	 * @return mixed
+	 * @return array An array of data for the table.
 	 */
-	public static function get_locations( $per_page = 10, $page_number = 1 ) {
+	public function get_locations() {
 
 		global $wpdb;
 
@@ -56,14 +53,6 @@ class Form_Locations_Table extends WP_List_Table {
 			$sql .= ! empty( $_REQUEST['order'] ) ? ' ' . esc_sql( $order ) : ' ASC';
 		}
 
-		if ( isset( $per_page ) ) {
-
-			$sql .= " LIMIT $per_page";
-
-			$sql .= ' OFFSET ' . ( $page_number - 1 ) * $per_page;
-
-		}
-
 		$result = $wpdb->get_results( $sql, 'ARRAY_A' ); //phpcs:ignore
 
 		return $result;
@@ -79,6 +68,11 @@ class Form_Locations_Table extends WP_List_Table {
 	function column_form_id( $item ) {
 
 		$form = GFAPI::get_form( $item['form_id'] );
+
+		// Form does not exist.
+		if ( ! $form ) {
+			return '<span style="color:red">Invalid Form ID Used</span>';
+		}
 
 		return '<a href="?page=gf_edit_forms&id=' . $form['id'] . '">' . $form['title'] . '</a>';
 	}
@@ -178,9 +172,23 @@ class Form_Locations_Table extends WP_List_Table {
 	 */
 	function prepare_items() {
 		$columns               = $this->get_columns();
+		$sortable              = $this->get_sortable_columns();
 		$hidden                = array();
-		$this->_column_headers = array( $columns, $hidden );
-		$this->items           = self::get_locations();
+		$this->_column_headers = array( $columns, $hidden, $sortable );
+
+		$per_page     = 25;
+		$current_page = $this->get_pagenum();
+		$data         = $this->get_locations( $per_page, $current_page );
+		$total_items  = count( $data );
+
+		$this->set_pagination_args(
+			array(
+				'total_items' => $total_items,
+				'per_page'    => $per_page,
+			)
+		);
+
+		$this->items = array_slice( $data, ( ( $current_page - 1 ) * $per_page ), $per_page );
 	}
 
 }