+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog #
 
+### 1.2.0 ###
+* Fixed a bug that caused some forms to be missed when scanning.
+* Fixed bugs related to duplicated form locations when editing posts.
+* Added support for multiple forms on the same post.
+* Added sortable columns to the location table.
+* Added pagination to the location table.
+
+### 1.0.2 ###
+* Fix bug causing plugin to crash on PHP 7
+
 ### 1.0.1 ###
 * Restructure Code Base
 * Fixed error where form location relations were duplicated